Refactor send assets flow to not redirect upon using SlidingSegmentControl
